Deduplicate redundant code and create `getJsonFiles.js`
